requirements.txt
# --- INTERFACE & VISUALISATION ---
streamlit>=1.30.0        # Framework cockpit (version stable pour Mac/Linux)
plotly>=5.18.0           # Graphiques dynamiques (Camembert/RUL)

# --- IA, DATA & MATHS ---
pandas>=2.1.0            # Manipulation des données
numpy>=1.24.0            # Calculs numériques haute performance
scikit-learn>=1.3.0      # Modèles RandomForest (Classification/Régression)
joblib>=1.3.0            # Chargement des modèles .pkl
scipy>=1.10.0            # Dépendance nécessaire pour scikit-learn

# --- CI/CD & QUALITÉ LOGICIELLE ---
pytest>=7.4.0            # Tests automatisés (déclenché par main.py)
pytest-cov>=4.1.0        # Mesure de la couverture de test (Bonus MLOps)

# --- OPÉRATIONS, LOGS & MONITORING ---
requests>=2.31.0         # Gestion des appels HTTP (Monitoring/API)
python-json-logger>=2.0.0 # Logs structurés pour le diagnostic d'incidents
prometheus-client>=0.17.0 # Export des métriques de performance
watchdog>=3.0.0          # Optimisation du rechargement Streamlit

# --- COMPATIBILITÉ SYSTÈME ---
setuptools               # Gestion de l'installation des paquets

fastapi 
uvicorn 
python-jose 
passlib==1.7.4
bcrypt==4.0.1 
sqlalchemy 
prometheus-client